bugfixes in frontend and backend for user management, code cleaning

+++ backend/src/main/java/hs/aalen/financial_assets_portfolio/service/AccountService.java
@@ -2,24 +2,19 @@ package hs.aalen.financial_assets_portfolio.service;
 
 import hs.aalen.financial_assets_portfolio.data.AccountDTO;
 import hs.aalen.financial_assets_portfolio.data.ExceptionDTO;
-import hs.aalen.financial_assets_portfolio.data.PurchaseDTO;
-import hs.aalen.financial_assets_portfolio.data.ShareDTO;
 import hs.aalen.financial_assets_portfolio.domain.Account;
 import hs.aalen.financial_assets_portfolio.exceptions.FormNotValidException;
 import hs.aalen.financial_assets_portfolio.persistence.AccountRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
-import java.util.List;
 import java.util.NoSuchElementException;
-import java.util.Optional;
 
 @Service
+
 public class AccountService {
 
-    public static final int STRING_MAX_LENGTH = 255;
+    public static final int STRING_MAX_LENGTH = 30;
 
     private final AccountRepository accountRepository;
 
@@ -44,11 +39,11 @@ public class AccountService {
         if (accountList.isEmpty()) {
             throw new NoSuchElementException();
         }
-        return new ArrayList<AccountDTO>(accountList.stream().map(AccountDTO::new).toList());
+        return new ArrayList<>(accountList.stream().map(AccountDTO::new).toList());
     }
 
     public void addAccount(String username, AccountDTO accountDTO) throws FormNotValidException {
-        ArrayList<ExceptionDTO> exceptionDTOList = this.validateForm(accountDTO);
+        ArrayList<ExceptionDTO> exceptionDTOList = this.validateAddForm(accountDTO);
         if (exceptionDTOList.isEmpty()){
             Account account = new Account (accountDTO);
             account.setRole(account.getRole().toUpperCase());
@@ -57,32 +52,47 @@ public class AccountService {
             throw new FormNotValidException("Formfehler", exceptionDTOList);
         }
     }
-
+    @Transactional
     public void updateAccount(String username, AccountDTO accountDTO) throws FormNotValidException{
         Account existingAccount = accountRepository.findByUsernameIgnoreCase(username);
-        ArrayList<ExceptionDTO> exceptionDTOList = this.validateForm(accountDTO);
+        ArrayList<ExceptionDTO> exceptionDTOList = this.validateUpdateForm(username, accountDTO);
+
         if (exceptionDTOList.isEmpty()){
-            existingAccount.setUsername(accountDTO.getUsername());
-            existingAccount.setName(accountDTO.getName());
-            existingAccount.setRole(accountDTO.getRole().toUpperCase());
-            existingAccount.setPassword(accountDTO.getPassword());
+            if(!existingAccount.getUsername().equals(accountDTO.getUsername()) && !accountDTO.getUsername().isEmpty()){
+                existingAccount = new Account(
+                        accountDTO.getUsername(),
+                        existingAccount.getPassword(),
+                        existingAccount.getName(),
+                        existingAccount.getRole());
+                accountRepository.deleteByUsername(username);
+            }
+            if(!existingAccount.getName().equals(accountDTO.getName()) && !accountDTO.getName().isEmpty()){
+                existingAccount.setName(accountDTO.getName());
+            }
+            if(!existingAccount.getPassword().equals(accountDTO.getPassword()) && !accountDTO.getPassword().isEmpty()){
+                existingAccount.setPassword(accountDTO.getPassword());
+            }
+            if(!existingAccount.getRole().equalsIgnoreCase(accountDTO.getRole()) && !accountDTO.getRole().isEmpty()){
+                existingAccount.setRole(accountDTO.getRole().toUpperCase());
+            }
             saveAccount(existingAccount);
         }else {
             throw new FormNotValidException("Formfehler", exceptionDTOList);
         }
     }
-
-    public void deleteAccountByUsername(String username) {
-        accountRepository.deleteByUsername(username);
+    @Transactional
+    public void deleteAccountByUsername(String username, String authenticatedUsername) {
+        if(!username.equals(authenticatedUsername)){
+            accountRepository.deleteByUsername(username);
+        }
     }
 
     public boolean checkAccountExists(String username) {
         Account account = accountRepository.findByUsernameIgnoreCase(username);
-
         return account != null && account.getUsername().equalsIgnoreCase(username);
     }
 
-    public ArrayList<ExceptionDTO> validateForm(AccountDTO accountDTO){
+    public ArrayList<ExceptionDTO> validateAddForm(AccountDTO accountDTO){
         ArrayList<ExceptionDTO> exceptions = new ArrayList<>();
 
         if (checkAccountExists(accountDTO.getUsername())){
@@ -108,9 +118,9 @@ public class AccountService {
             exceptions.add(new ExceptionDTO(
                     "name", "Der Name darf nicht länger als 255 Zeichen sein"));
         }
-        if(accountDTO.getRole().length() > STRING_MAX_LENGTH){
+        if(!(accountDTO.getRole().equals("ADMIN") || accountDTO.getRole().equals("USER"))){
             exceptions.add(new ExceptionDTO(
-                    "role", "Die Rolle darf nicht länger als 255 Zeichen sein"));
+                    "role", "Rolle nicht valide"));
         }
         if(accountDTO.getPassword().length() > STRING_MAX_LENGTH){
             exceptions.add(new ExceptionDTO(
@@ -118,4 +128,43 @@ public class AccountService {
         }
         return exceptions;
     }
+
+    public ArrayList<ExceptionDTO> validateUpdateForm(String initUsername, AccountDTO accountDTO){
+        Account account = accountRepository.findByUsernameIgnoreCase(initUsername);
+        ArrayList<ExceptionDTO> exceptions = new ArrayList<>();
+
+        if (!(accountDTO.getUsername().equalsIgnoreCase(account.getUsername())) && !accountDTO.getUsername().isEmpty()){
+            if (checkAccountExists(accountDTO.getUsername())){
+                exceptions.add(new ExceptionDTO("username", "Account mit diesem Benutzernamen bereits vorhanden"));
+            }
+            if((accountDTO.getUsername() == null || accountDTO.getUsername().isEmpty())){
+                exceptions.add(new ExceptionDTO("username", "Bitte füllen Sie den Benutzername aus"));
+            }
+            if(accountDTO.getUsername().length() > STRING_MAX_LENGTH){
+                exceptions.add(new ExceptionDTO(
+                        "username", "Der Benutzername darf nicht länger als 30 Zeichen sein"));
+            }
+        }
+
+        if(!(accountDTO.getName().equalsIgnoreCase(account.getName())) && !accountDTO.getName().isEmpty()){
+            if(accountDTO.getName().length() > STRING_MAX_LENGTH){
+                exceptions.add(new ExceptionDTO(
+                        "name", "Der Name darf nicht länger als 30 Zeichen sein"));
+            }
+        }
+
+        if(! (accountDTO.getRole().equals("ADMIN") || accountDTO.getRole().equals("USER"))){
+            exceptions.add(new ExceptionDTO(
+                    "role", "Rolle nicht valide"));
+        }
+
+        if(!(accountDTO.getPassword().equalsIgnoreCase(account.getPassword())) && !accountDTO.getPassword().isEmpty()){
+            if(accountDTO.getPassword().length() > STRING_MAX_LENGTH){
+                exceptions.add(new ExceptionDTO(
+                        "password", "Das Passwort darf nicht länger als 30 Zeichen sein"));
+            }
+        }
+        return exceptions;
+    }
+
 }
